Asoc: msm8x16: fix compilation for qrd
support internal codec on qrd and external codec on customer target.

+++ sound/soc/codecs/wsa881x-temp-sensor.h
@@ -35,13 +35,14 @@ struct wsa881x_tz_priv {
 
 #ifdef CONFIG_SND_SOC_WSA881X_SENSORS
 
-int wsa881x_get_temp(struct thermal_zone_device *tz_dev, int *temp);
+int wsa881x_get_temp(struct thermal_zone_device *tz_dev, unsigned long *temp);
 int wsa881x_init_thermal(struct wsa881x_tz_priv *tz_pdata);
 void wsa881x_deinit_thermal(struct thermal_zone_device *tz_dev);
 
 #else
 
-int wsa881x_get_temp(struct thermal_zone_device *tz_dev, int *temp)
+int wsa881x_get_temp(struct thermal_zone_device *tz_dev,
+		unsigned long *temp)
 {
 	return 0;
 }
